# Pool timers to avoid allocations