Requests are somewhat allocation heavy. This should help a bit. Also has a small positive effect on requests performance (more for the old request style).

Benchstat (base 9147cfac69833916f9c73353ba43f939815c330a):

```
name                              old time/op    new time/op    delta
Request-8                           43.9µs ± 5%    43.4µs ± 1%     ~     (p=0.690 n=5+5)
OldRequest-8                        60.0µs ± 3%    57.5µs ± 2%   -4.30%  (p=0.008 n=5+5)

name                              old alloc/op   new alloc/op   delta
Request-8                             885B ± 0%      693B ± 0%  -21.71%  (p=0.008 n=5+5)
OldRequest-8                        2.63kB ± 0%    2.44kB ± 0%   -7.27%  (p=0.000 n=5+4)

name                              old allocs/op  new allocs/op  delta
Request-8                             17.0 ± 0%      14.0 ± 0%  -17.65%  (p=0.008 n=5+5)
OldRequest-8                          55.0 ± 0%      52.0 ± 0%   -5.45%  (p=0.008 n=5+5)
```

Other benchmarks show no changes.