@nussjustin 2017-05-31
Test failure is unrelated and also happens on master (though only 2 times in 20 local runs, 10 with and 10 without -race).