@nussjustin 2017-05-31
Since this works with sync.Pool testing is somewhat tricky. The test would need to disable GC (and reenable afterwards) and set GOMAXPROCS to 1 (and reset it afterwards) to prevent the pool from being emptied and to avoid rescheduling onto another P.

If you want I can push a test.