@tylertreat 2017-05-31
lgtm. Would be nice to have some basic tests around `timerPool`, though I agree it's a bit tricky to test with the GC behavior.