@nussjustin 2017-05-31
Pushed a new commit with a test that should work.