@nussjustin 2017-05-31
Ah, okay. sync.Pool will randomly drop values when compiling with -race to prevent assuming it's behaviour. I changed the test to just check the behaviour of the timer instead.